refactor: external 패키지 분리
