+++ src/main/java/roomescape/reservation/external/dto/respone/PaymentApproveResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto.response;
+package roomescape.reservation.external.dto.respone;
 
 public record PaymentApproveResponse(
         String paymentKey,
