+++ src/main/java/roomescape/reservation/service/PaymentService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.reservation.entity.Payment;
+import roomescape.reservation.external.client.PaymentRestClient;
 import roomescape.reservation.repository.PaymentRepository;
 
 @Service
