+++ src/test/java/roomescape/payment/PaymentRestClientTest.java
@@ -15,10 +15,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestClient;
-import roomescape.reservation.dto.response.PaymentApproveResponse;
-import roomescape.reservation.error.exception.PaymentClientException;
-import roomescape.reservation.error.handler.PaymentResponseErrorHandler;
-import roomescape.reservation.service.PaymentRestClient;
+import roomescape.reservation.external.client.PaymentRestClient;
+import roomescape.reservation.external.dto.respone.PaymentApproveResponse;
+import roomescape.reservation.external.error.exception.PaymentClientException;
+import roomescape.reservation.external.error.handler.PaymentResponseErrorHandler;
 
 class PaymentRestClientTest {
